+++ changelog.md
@@ -9,6 +9,7 @@
 * Made it possible to order the groups by drag and drop.
 * Made it possible to order the applications by drag and drop.
 * Made it possible to order the tables by drag and drop.
+* Added today, this month and this year filter.
 
 ## Released (2021-05-11)